Judgement Type component
Creates a judgement type component that shows a green (solved), red (penalty), or
cyan (neither) badge with the judgement type's acronym. Uses it on the team's list
of submissions, as well as the Problem component.

I'm still not sure if the components should have fixed width (this one does) or
fill the parent (like Problem), but this is prettier in the meantime and easy to
change later.