Add scoreboard row to team page
Added a 'mode' option to the scoreboard row that shows a 'summary', which basically
means 'no logo or team name'. Added this row to the top of every team page. Added a
light problem label in this mode b/c it wasn't as obvious when you're not looking at
the whole scoreboard.

We'll need to improve the formatting and layout later to make it look a bit prettier
and easier to understand, but this fixes the gap by providing the team's current
standing on every team page.

Fixes #28.