Team area map
Adds a FloorMap component (an HTML canvas) and a new page /map, not linked from
anywhere else for now since it is 'beta' and missing some features [*]. It shows
the contest floor if one exists for the contest with all the team desks. Hovering
highlights a team [*], and clicking on them redirects to the team's page.

Contest types are updated to support the optional map-info and team location.
FloorMap is a reusable component.

[*] Remaining issues:
- it's a bit plain, could use a bit more design.
- desks rotate, but the selection area doesn't - need to switch to shapes or
  rotated rects instead.
- font doesn't resize based on canvas size.
- component could use more features, e.g. show the logo on the selected team.
- onclick should just be an event, and let the client decide if it does anything.

Majority of #43.