Fix layout and resizing issues
Fix the layout so that the banner is always across the bottom of the screen and each
page decides if it should fill the remaining content area, scroll, or not.

Homepage, team info page, and scoreboard are set to add scrollbars as necessary.
'Fluid' disabled for Video so that it doesn't overflow. All video set to fit
within the current screen size with black edges as necessary to fill.

Fixes #30.