Improve tag support and use light images
Update the 'best logo' support to match the Java ICPC tools and look for logos in
order by matching tag, non-opposite tag, or remaining images.

Use logo tags hard-coded to the background they're on for now. When we correctly
support light/dark mode correctly we'll need to adapt these.

Fixes #74.