Handle reload error without crashing
Today, no contest data is preloaded. We could do a quick cache of everything (and
event feed support is in progress...), but for now the first person to hit any page
requiring data triggers the load just of that part of the contest; after a few pages
or people everything is cached and fast.

If the Contest API it is connected to or the network fails during this period it's
not a huge deal; the one person will get an error but it should retry for the next
person until it succeeds, then everything becomes cached and really fast.

Every 5s the data that tends to change (submissions, judgements, scoreboard) is
force-reloaded, and if that fails it causes the crash in #75. This just puts a catch
block around it so that there will be a log message, and it should try again in 5s.

Fixes #75.